Making it modular with express routes
